+++ src/extension-test/integration/suite/index.ts
@@ -12,7 +12,7 @@ export function run(): Promise<void> {
       "reporterEnabled": "mocha-junit-reporter, spec",
       "mochaJunitReporterReporterOptions": {
         "mochaFile": "junit/test-results.xml"
-    }
+      }
     }
   });
   mocha.useColors(true);
@@ -20,11 +20,10 @@ export function run(): Promise<void> {
   const testsRoot = path.resolve(__dirname, '..');
 
   return new Promise((c, e) => {
-    glob('**/**.test.js', { cwd: testsRoot }, (err, files) => {
+    glob('**/**-test.js', { cwd: testsRoot }, (err, files) => {
       if (err) {
         return e(err);
       }
-
       // Add files to the test suite
       files.forEach(f => mocha.addFile(path.resolve(testsRoot, f)));